// src/core/api/generated/spring/schemas/SystemConfigDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */
import type { SystemConfigDTOSystemConfigType } from './SystemConfigDTOSystemConfigType';
import type { SystemConfigDTOStatus } from './SystemConfigDTOStatus';

export interface SystemConfigDTO {
  id?: number;
  /**
   * @minLength 2
   * @maxLength 100
   * @pattern ^[A-Za-z0-9_.:-]+$
   */
  configKey: string;
  systemConfigType: SystemConfigDTOSystemConfigType;
  /**
   * @minLength 0
   * @maxLength 255
   */
  description?: string;
  status: SystemConfigDTOStatus;
  /**
   * @minLength 0
   * @maxLength 50
   */
  createdBy?: string;
  createdDate?: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  lastModifiedBy?: string;
  lastModifiedDate?: string;
}